pyproject.toml
[project]
name = "wuwa-wiki-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "fastapi[standard]>=0.115.12",
    "httpx>=0.28.1",
    "strawberry-graphql[fastapi]>=0.266.0",
]

[dependency-groups]
dev = [
    "ruff>=0.11.7",
]
